+++ src/main/java/org/anjeyy/soba/common/CustomStyleSheet.java
@@ -1,0 +1,10 @@
+package org.anjeyy.soba.common;
+
+import javafx.scene.Node;
+
+public interface CustomStyleSheet {
+
+    default void add(Node node, String cssClassName) {
+        node.getStyleClass().add(cssClassName);
+    }
+}
