+++ src/main/java/org/anjeyy/soba/welcome/WelcomeController.java
@@ -3,15 +3,10 @@ package org.anjeyy.soba.welcome;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXButton.ButtonType;
 import com.jfoenix.controls.JFXClippedPane;
-import java.io.InputStream;
-import java.util.Objects;
 import java.util.Random;
 import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.StackPane;
-import javafx.stage.Stage;
+import javafx.scene.layout.BorderPane;
+import javafx.scene.layout.Pane;
 
 public class WelcomeController {
 
@@ -20,11 +15,7 @@ public class WelcomeController {
     //todo set as bottom half
 
 
-    public StackPane setup(Stage currentStage, Node followingNode) {
-        InputStream inputStream = Objects.requireNonNull(
-            getClass().getResourceAsStream("/image/welcome-background.jpg"), "Could not load image.");
-        Image backgroundImage = new Image(inputStream);
-
+    public BorderPane setup() {
         JFXButton button = new JFXButton("JFX foenix button");
         button.setStyle(
             "-fx-background-color:rgb(" +
@@ -35,14 +26,15 @@ public class WelcomeController {
         button.setButtonType(ButtonType.FLAT);
         button.setAlignment(Pos.CENTER);
         button.getStyleClass().add("button");
-//        button.setOnAction(e -> currentStage.setScene(followingNode));
+//        StackPane stackPane = new StackPane(button);
+        JFXClippedPane stackPane = new JFXClippedPane(button);
 
-        ImageView backgroundImageView = new ImageView(backgroundImage);
-//        backgroundImageView.setOnMouseClicked(e -> currentStage.setScene(followingScene));
-//        StackPane stackPane = new StackPane(backgroundImageView, button);
-        JFXClippedPane stackPane = new JFXClippedPane();
-        stackPane.setMaxSize(1000, 1000);
+        BorderPane borderPane = new BorderPane();
+        borderPane.setCenter(new Pane());
+//        borderPane.setCenter(button);
+
         stackPane.getStyleClass().add("entry-screen");
-        return stackPane;
+        borderPane.getStyleClass().add("entry-screen");
+        return borderPane;
     }
 }
