+++ src/main/java/org/anjeyy/soba/welcome/WelcomeView.java
@@ -1,4 +1,14 @@
 package org.anjeyy.soba.welcome;
 
-public class WelcomeView {
+import javafx.scene.Parent;
+import org.anjeyy.soba.common.MainView;
+
+public class WelcomeView implements MainView {
+
+
+    @Override
+    public Parent asParent() {
+        //todo - for main view
+        return null;
+    }
 }
