+++ src/test/java/org/anjeyy/soba/window/LeftWindowToolbarTest.java
@@ -3,6 +3,7 @@ package org.anjeyy.soba.window;
 import com.jfoenix.controls.JFXButton;
 import javafx.scene.Node;
 import javafx.scene.control.Labeled;
+import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -21,7 +22,7 @@ class LeftWindowToolbarTest {
         WindowToolbarModel windowToolbarModel = new WindowToolbarModel();
         WindowToolbarController windowToolbarController = new WindowToolbarController(windowToolbarModel);
         WindowToolbarView uut = new WindowToolbarView(windowToolbarController, windowToolbarModel);
-        stage.setScene(uut.setup());
+        stage.setScene(uut.setup(new Pane()));
         stage.show();
     }
