+++ src/test/java/org/anjeyy/soba/window/MinimizableWindowToolbarTest.java
@@ -1,6 +1,7 @@
 package org.anjeyy.soba.window;
 
 import javafx.beans.property.ReadOnlyBooleanProperty;
+import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Disabled;
@@ -20,7 +21,7 @@ class MinimizableWindowToolbarTest {
         WindowToolbarModel windowToolbarModel = new WindowToolbarModel();
         WindowToolbarController windowToolbarController = new WindowToolbarController(windowToolbarModel);
         WindowToolbarView uut = new WindowToolbarView(windowToolbarController, windowToolbarModel);
-        stage.setScene(uut.setup());
+        stage.setScene(uut.setup(new Pane()));
         stage.show();
     }
